replace frames with new ones
main reason behind this was that the copy right of the old frames was unclear.
The new frames are based on freely availabel images from http://www.tempees.com/
With this change the dimension of the mobile phone screens have been updated to the
official specs of the corresponding phones.